JWT validation, Authorization for CRUD

+++ restapi-project/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -46,6 +47,7 @@ exports.createPost = (req, res, next) => {
         throw err;
     }
 
+    let creator;
     const imageUrl = req.file.path.replace('\\', '/');
     const title = req.body.title;
     const content = req.body.content;
@@ -53,16 +55,26 @@ exports.createPost = (req, res, next) => {
         title,
         imageUrl,
         content,
-        creator: {
-            name: 'Avnish'
-        }
+        creator: req.userId
     });
 
     post.save()
-        .then(result => {
+        .then(() => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(() => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             });
         })
         .catch(err => {
@@ -127,6 +139,12 @@ exports.updatePost = (req, res, next) => {
                 throw err;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw err;
+            }
+
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -161,10 +179,23 @@ exports.deletePost = (req, res, next) => {
                 throw err;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw err;
+            }
+
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
-        .then(result => {
+        .then(() => {
+            return User.findById(req.userId);
+        })
+        .then(user=>{
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(()=>{
             res.status(200).json({ message: 'Post deleted' });
         })
         .catch(err => {
